+++ threedigrid/__init__.py
@@ -12,4 +12,4 @@ except ImportError:
 
 # Threedigrid version number is automatic updated with zest.releaser
 # the version number in setup.py is updated using the find_version()
-__version__ = '0.2.8.dev0'
+__version__ = '0.2.9.dev0'