+++ threedigrid/admin/gridresultadmin.py
@@ -55,7 +55,7 @@ class GridH5ResultAdmin(GridH5Admin):
     def set_timeseries_chunk_size(self, new_chunk_size):
         """
         overwrite the default chunk size for timeseries queries.
-        :param new_chunk_size <int> or <slice>: new chunk size for
+        :param new_chunk_size <int>: new chunk size for
             timeseries queries
         :raises ValueError when the given value is less than 1
         """
@@ -173,6 +173,10 @@ class GridH5ResultAdmin(GridH5Admin):
             logger.error(
                 'Attribute threedicore_version could not be found in result file')  # noqa
         return ''
+
+    def close(self):
+        super(GridH5ResultAdmin, self).close()
+        self.netcdf_file.close()
 
 
 class GridH5AggregateResultAdmin(GridH5ResultAdmin):