+++ threedigrid/orm/base/timeseries_mixin.py
@@ -57,7 +57,7 @@ class ResultMixin(object):
             return
 
         fields = {}
-
+        count = self.get_field_value('id').size
         for v, k in six.iteritems(self.Meta.subset_fields):
             _source_name = list(k.values())
             if not _source_name:
@@ -65,7 +65,7 @@ class ResultMixin(object):
             source_name = _source_name[0]
 
             fields[v] = TimeSeriesSubsetArrayField(
-                source_name=source_name, size=self.count
+                source_name=source_name, size=count
             )
 
         self._meta.add_fields(fields, hide_private=True)
@@ -171,10 +171,12 @@ class ResultMixin(object):
 
     @property
     def dt_timestamps(self):
-        return [t.isoformat() for t in num2date(
+        return [
+            t.isoformat() for t in num2date(
                 self.timestamps,
-                units=self._datasource.get('time').getncattr('units')
-            )]
+                units=self._datasource.get('time').getncattr('units'))
+        ]
+
 
 class AggregateResultMixin(ResultMixin):
     """